Rename hightlight.ts to highlight.ts (Fix typo), rename references as well
