+++ src/app/services/highlight.service.ts
@@ -18,7 +18,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { API_CONFIG } from '../config/api-config';
-import { Highlight } from '../models/domain/hightlight';
+import { Highlight } from '../models/domain/highlight';
 import { HighlightAdd } from '../models/resources/highlight-add';
 import { HighlightUpdate } from '../models/resources/highlight-update';
 import { HttpBaseService } from './http-base.service';
